+++ tests/models/test_privacy_request.py
@@ -8,7 +8,8 @@ from pydantic import ValidationError
 from sqlalchemy.orm import Session
 
 from fidesops.common_exceptions import ClientUnsuccessfulException, PrivacyRequestPaused
-from fidesops.models.policy import Policy
+from fidesops.graph.config import CollectionAddress
+from fidesops.models.policy import ActionType, Policy
 from fidesops.models.privacy_request import PrivacyRequest, PrivacyRequestStatus
 from fidesops.schemas.redis_cache import PrivacyRequestIdentity
 from fidesops.util.cache import FidesopsRedis, get_identity_cache_key
@@ -357,3 +358,57 @@ class TestPrivacyRequestTriggerWebhooks:
             )
             with pytest.raises(ValidationError):
                 privacy_request.trigger_policy_webhook(webhook)
+
+
+class TestCachePausedLocation:
+    def test_privacy_request_cache_paused_location(self, privacy_request):
+        assert privacy_request.get_paused_step_and_collection() == (None, None)
+
+        paused_step = ActionType.erasure
+        paused_location = CollectionAddress("test_dataset", "test_collection")
+        privacy_request.cache_paused_step_and_collection(paused_step, paused_location)
+
+        assert privacy_request.get_paused_step_and_collection() == (
+            paused_step,
+            paused_location,
+        )
+
+    def test_privacy_request_unpause(self, privacy_request):
+        privacy_request.cache_paused_step_and_collection()
+
+        assert privacy_request.get_paused_step_and_collection() == (None, None)
+
+
+class TestCacheManualInput:
+    def test_cache_manual_input(self, privacy_request):
+        paused_location = CollectionAddress("test_dataset", "test_collection")
+        manual_data = [{"id": 1, "name": "Jane"}, {"id": 2, "name": "Hank"}]
+
+        privacy_request.cache_manual_input(paused_location, manual_data)
+
+        assert privacy_request.get_manual_input(paused_location,) == {
+            f"EN_MANUAL_INPUT__{privacy_request.id}__{paused_location.value}": manual_data
+        }
+
+    def test_cache_empty_manual_input(self, privacy_request):
+        paused_location = CollectionAddress("test_dataset", "test_collection")
+        manual_data = []
+
+        privacy_request.cache_manual_input(paused_location, manual_data)
+
+        assert (
+            privacy_request.get_manual_input(
+                paused_location,
+            )
+            == {f"EN_MANUAL_INPUT__{privacy_request.id}__{paused_location.value}": []}
+        )
+
+    def test_no_manual_data_in_cache(self, privacy_request):
+        paused_location = CollectionAddress("test_dataset", "test_collection")
+
+        assert (
+            privacy_request.get_manual_input(
+                paused_location,
+            )
+            == {}
+        )